apps/ml-scoring/pyproject.toml
[tool.poetry]
name = "finrisk-ml-scoring"
version = "1.0.0"
description = "FinRisk AI Copilot - ML Scoring Service with SHAP explainability"
authors = ["Manikandan Bala <manikandan@finrisk.ai>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"

# Machine Learning (core only for now)
scikit-learn = "^1.3.2"
pandas = "^2.1.4"
numpy = "^1.25.2"
# scipy = "^1.11.4"
# xgboost = "^2.0.2"
# lightgbm = "^4.1.0"
# catboost = "^1.2.2"

# Explainability (simplified for now)
# shap = "^0.43.0"
# lime = "^0.2.0.1"
# interpret = "^0.4.3"

# Feature Store (commented out due to compatibility issues)
# feast = "^0.35.0"

# Model Serving
onnxruntime = "^1.16.3"
onnx = "^1.15.0"
# mlflow = "^2.8.1"  # commented out due to dependency conflicts

# Database
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
alembic = "^1.13.1"

# Cache and Message Queue
redis = "^5.0.1"
aiokafka = "^0.10.0"

# Monitoring and Observability
prometheus-client = "^0.19.0"
opentelemetry-api = "^1.21.0"
opentelemetry-sdk = "^1.21.0"
opentelemetry-instrumentation-fastapi = "^0.42b0"

# Utilities
pydantic-settings = "^2.1.0"
structlog = "^23.2.0"
httpx = "^0.25.2"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
tenacity = "^8.2.3"
click = "^8.1.7"
joblib = "^1.3.2"
dill = "^0.3.7"

# Data Processing
pyarrow = "^14.0.1"
polars = "^0.19.19"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.6.0"
jupyter = "^1.0.0"
matplotlib = "^3.8.2"
seaborn = "^0.13.0"

[tool.poetry.group.test.dependencies]
testcontainers = "^3.7.1"
factory-boy = "^3.3.0"
faker = "^20.1.0"
pytest-benchmark = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
